restructure whipser/local
