+++ mobile/examples/whisper/local/android/app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.SpeechRecognition"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="ai.onnxruntime.example.whisperLocal.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
