+++ mobile/examples/whisper/local/android/app/src/main/java/ai/onnxruntime/example/whisperLocal/MainActivity.kt
@@ -1,5 +1,6 @@
-package ai.onnxruntime.example.speechrecognition
+package ai.onnxruntime.example.whisperLocal
 
+import ai.onnxruntime.example.speechrecognition.R
 import android.Manifest
 import android.content.pm.PackageManager
 import androidx.appcompat.app.AppCompatActivity
