+++ mobile/examples/whisper/local/android/app/src/main/java/ai/onnxruntime/example/whisperLocal/TensorUtils.kt
@@ -1,4 +1,4 @@
-package ai.onnxruntime.example.speechrecognition
+package ai.onnxruntime.example.whisperLocal
 
 import ai.onnxruntime.OnnxTensor
 import ai.onnxruntime.OrtEnvironment
