+++ mobile/examples/whisper/local/android/app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context="ai.onnxruntime.example.whisperLocal.MainActivity">
 
     <ImageView
         android:id="@+id/logo_image_view"
